+++ tests/testthat/test-modifiers.R
@@ -37,12 +37,6 @@ test_that("subsetting preserves class and options", {
   expect_equal(x[], x)
 })
 
-test_that("useful error message for NA_character_", {
-  expect_snapshot(error = TRUE, {
-    type(NA_character_)
-  })
-})
-
 test_that("useful error message for vector that includes NAs", {
   expect_snapshot(error = TRUE, {
     type(c("a", "b", NA_character_, "c"))