+++ R/modifiers.R
@@ -202,6 +202,13 @@ type.stringr_fixed <- function(x, error_call = caller_env()) {
 }
 #' @export
 type.character <- function(x, error_call = caller_env()) {
+  if (any(is.na(x))) {
+    cli::cli_abort(
+      tr_("{.arg pattern} must be a character vector that does not contain NAs."),
+      call = error_call
+    )
+  }
+
   if (identical(x, "")) "empty" else "regex"
 }
 
@@ -213,7 +220,7 @@ type.default <- function(x, error_call = caller_env()) {
   }
 
   cli::cli_abort(
-    tr_("{.arg pattern} must be a string, not {.obj_type_friendly {x}}."),
+    tr_("{.arg pattern} must be a character vector, not {.obj_type_friendly {x}}."),
     call = error_call
   )
 }