+++ tests/testthat/test-modifiers.R
@@ -37,6 +37,13 @@ test_that("subsetting preserves class and options", {
   expect_equal(x[], x)
 })
 
+test_that("useful errors for NAs", {
+  expect_snapshot(error = TRUE, {
+    type(NA)
+    type(c("a", "b", NA_character_, "c"))
+  })
+})
+
 test_that("stringr_pattern methods", {
   ex <- coll(c("foo", "bar"))
   expect_true(inherits(ex[1], "stringr_pattern"))