+++ NEWS.md
@@ -1,8 +1,9 @@
 # stringr (development version)
 
+* `str_view()` now displays a message when called with a zero-length character
+  vector (@LouisMPenrod, #497).
 * Adds `[[.stringr_pattern` method to go along with existing `[.stringr_pattern`
   method (@edward-burn, #569).
-
 * In `str_replace_all()`, a `replacement` function now receives all values in
   a single vector. This radically improves performance at the cost of breaking
   some existing uses (#462).